app/services/search_service.py
import logging
from typing import List, Dict, Any
from google.generativeai import genai
import os
from app.config.settings import settings

logger = logging.getLogger(__name__)

class SearchService:
    def __init__(self):
        # We can use our search_web tool or a simple LLM-based search if we had access.
        # Since I am the agent, I can provide a way for the AI to "think" about search.
        pass

    async def get_latest_tech_info(self, query: str) -> str:
        """
        In a real professional app, this would use Google Search API or Brave Search.
        For this project, we'll simulate high-level knowledge retrieval.
        """
        logger.info(f"Retrieving latest tech info for: {query}")
        # Logic to fetch latest tech trends would go here
        return f"Latest info for {query}: Use Next.js 14 App Router, Tailwind v3+, and FastAPI Pydantic v2."

search_service = SearchService()
